+++ RELEASE-NOTES.md
@@ -40,7 +40,6 @@ END TEMPLATE-->
 ### New features
 
 * `TextEdit.OnTextChanged`
-* Add ICollection<T> versions of Pick and PickAndTake for System.Random.
 
 ### Bugfixes
