+++ Robust.Shared/Random/IRobustRandom.cs
@@ -101,3 +101,69 @@ public interface IRobustRandom
         return (byte) Next(minValue, maxValue);
     }
 }
+
+public static class RandomHelpers
+{
+    public static void Shuffle<T>(this System.Random random, IList<T> list)
+    {
+        var n = list.Count;
+        while (n > 1)
+        {
+            n -= 1;
+            var k = random.Next(n + 1);
+            (list[k], list[n]) = (list[n], list[k]);
+        }
+    }
+
+    public static bool Prob(this System.Random random, double chance)
+    {
+        return random.NextDouble() < chance;
+    }
+
+    /// <summary>
+    /// Picks a random element from a set and returns it.
+    /// </summary>
+    public static T Pick<T>(this System.Random random, HashSet<T> set)
+    {
+        var index = random.Next(set.Count);
+        var idx = 0;
+
+        foreach (var tile in set)
+        {
+            if (idx == index)
+                return tile;
+
+            idx++;
+        }
+
+        throw new InvalidOperationException();
+    }
+
+    /// <summary>
+    /// Picks a random from a set then removes it and returns it.
+    /// </summary>
+    public static T Take<T>(this System.Random random, HashSet<T> set)
+    {
+        var tile = Pick(random, set);
+        set.Remove(tile);
+        return tile;
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static byte NextByte(this System.Random random, byte maxValue)
+    {
+        return NextByte(random, 0, maxValue);
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static byte NextByte(this System.Random random)
+    {
+        return NextByte(random, byte.MaxValue);
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static byte NextByte(this System.Random random, byte minValue, byte maxValue)
+    {
+        return (byte) random.Next(minValue, maxValue);
+    }
+}
